Move flask-based tests of providers manager to FAB provider tests
When we remove flask from airlfow those tests in providers manager
fail, moving them to FAB which uses flask (and the features
tested are only anyhow working when FAB provider is installed) should
solve the issue.
